EV-114 Romove schemas file
